공통 코드 리팩토링.
